Fix test on slow machines, increasing the sleep time
The original sleep time is too short, the tests failed on
architectures like mips*, armhf. I think it's because the
python interpreter can't start in 0.1s.

After increasing the sleep time, it passed on most architectures
which Debian supports.

The result can be found at:
https://buildd.debian.org/status/logs.php?pkg=dumb-init
(compare the build result of 1.2.2-1.1 and 1.2.2-1)

Signed-off-by: Shengjing Zhu <zhsj@debian.org>