Use "%.*s" when printing VERSION
Commit 7bb2fef0e244220cc3f0f1dfc4317297f158b943 attempted to fix
the "missing NUL-terminator in VERSION" issue by replacing the "%s"
conversion specifier with "%*s". However, this fix itself was wrong,
as "%*s" specifies the field width, i.e. _minimum_ number of characters
to print. The proper solution is to use "%.*s", which specifies the
precision, i.e. the _maximum_ number of printed characters.